+++ src/GLWidget.h
@@ -22,7 +22,7 @@ public:
 	void cleanup();
 	void add_shape(std::string shape);
 	void select_entity(uint selected_id);
-	void move(uint id, char pos, uint value);
+	void move(uint id, char pos, float value);
 
 protected:
 	// Override from QOpenGLWidget
